% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptidy.R
\name{augment_kmeans}
\alias{augment_kmeans}
\title{Cluster Assignments of the Original Data Set}
\usage{
augment_kmeans(Model, X)
}
\arguments{
\item{Model}{A \code{kmeans} object created by \code{\link[stats:kmeans]{stats::kmeans()}}.}

\item{X}{data.frame of the original data set.}
}
\value{
data.frame of the original data and their cluster assignment.
}
\description{
Cluster Assignments of the Original Data Set
}
\examples{
library(tidyverse)
library(stats)
data(iris)
data <- iris \%>\% select(-Species)
kclust <- kmeans(data, centers = 3)
augment_kmeans(kclust, data)
}
