% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptidy.R
\name{tidy_kmeans}
\alias{tidy_kmeans}
\title{Tidy Output for KMeans Clustering}
\usage{
tidy_kmeans(Model, X)
}
\arguments{
\item{X}{data.frame of the original data set.}

\item{model}{A \code{kmeans} object created by \code{\link[stats:kmeans]{stats::kmeans()}}.}
}
\value{
tibble of infomration associated with each cluster
}
\description{
Creates a tidy dataframe containing information at the cluster level for a
kmeans clustering algorithm
}
\examples{
library(tidyverse)
library(tidymodels)
library(stats)
data(iris)
data <- iris \%>\% select(-Species)
kclust <- kmeans(data, centers = 3)
tidy_kmeans(kclust, data)
}
