% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptidy.R
\name{tidy_lr}
\alias{tidy_lr}
\title{Tidy Output for tidymodels Linear Regression}
\usage{
tidy_lr(model, X)
}
\value{
output data.frame
}
\description{
Create a tidy output for tidymodels linear_reg() object. The output dataframe
would have n+1 rows, where n is number of features and 3 columns describing feature
names, coefficients/intercept and p-values
}
\examples{
library(tidyverse)
library(tidymodels)
data("longley")
my_lr <- linear_reg() \%>\%
set_engine("lm") \%>\%
fit(Employed~., data = longley)
tidy_lr(my_lr,X)
}
