% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptidy.R
\name{augment_lr}
\alias{augment_lr}
\title{Augmented Output for Tidymodel's Linear Regression}
\usage{
augment_lr(my_lr, x, y)
}
\arguments{
\item{my_lr}{the linear model to augment}

\item{x}{the data frame containing the explanatory variables}

\item{y}{the data frame containing the target variable}
}
\value{
output data.frame
}
\description{
Adds two columns to the original data of the scikit learn's linear regression model. This includes predictions and residuals.
}
\examples{
# Import libraries
library(dplyr)
library(sptidy)
data("longley")
my_lr <- lm(Employed~., data = longley)
augment_lr(my_lr, (longley \%>\% select(!Employed)),
 as.data.frame(longley$Employed))
}
