% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptidy.R
\name{tidy_lr}
\alias{tidy_lr}
\title{Tidy Output for Linear Regression}
\usage{
tidy_lr(model)
}
\arguments{
\item{model}{lm_description}
}
\value{
output data.frame
}
\description{
Create a tidy output for lm() object. The output
would have n+1 rows, where n is number of features and
4 columns describing coefficient estimates, standard
error, t-statistics and p-values
}
\examples{

data("longley")
my_lr <- lm(Employed~., data = longley)
tidy_lr(my_lr)

}
